Feature/feed pipeline with tde (#90)
* Initial draft - Not yet tested

* Testing and tweaks. Terraform changes completed.

* Added support for upstream DDEI token and removed implementation of CGI Document Extraction Service until the specifics of it are known

* Latest fixes on testing TDE integration

* Ironed out some bugs in the evaluate documents process.

* Final bug fixes for document evaluation and brought the terraform up-to-date with the DDEI function app config definitions. Hardcoded function keys for now.

+++ coordinator.tests/Factories/EvaluateExistingDocumentsHttpRequestFactoryTests.cs
@@ -22,7 +22,7 @@ namespace coordinator.tests.Factories
 {
 	public class EvaluateExistingDocumentsHttpRequestFactoryTests
 	{
-        private readonly int _caseId;
+        private readonly long _caseId;
         private readonly List<CaseDocument> _existingCaseDocuments;
 		private readonly AccessToken _clientAccessToken;
 		private readonly string _content;
